add generic object
