+++ src/main/java/employee/service/EngineerService.java
@@ -1,0 +1,7 @@
+package employee.service;
+
+
+public interface EngineerService{
+
+
+}
