+++ src/main/java/employee/service/EngineerServiceImpl.java
@@ -1,0 +1,16 @@
+package employee.service;
+
+import employee.pojo.Engineer;
+
+public class EngineerServiceImpl implements BaseService<Engineer>   {
+
+    @Override
+    public void add(Engineer object) {
+
+    }
+
+    @Override
+    public void delete(Engineer object) {
+
+    }
+}
