+++ src/main/java/employee/service/WorkerService.java
@@ -1,0 +1,4 @@
+package employee.service;
+
+public interface WorkerService {
+}
