+++ src/main/java/employee/service/WorkerServiceImpl.java
@@ -1,0 +1,17 @@
+package employee.service;
+
+import employee.pojo.Engineer;
+import employee.pojo.Worker;
+
+public class WorkerServiceImpl implements BaseService<Worker> {
+
+    @Override
+    public void add(Worker object) {
+
+    }
+
+    @Override
+    public void delete(Worker object) {
+
+    }
+}
